feat: 블로그 글 헤더, 좋아요 컴포넌트 분리
